+++ app/providers/services/stats/Provider.js
@@ -7,8 +7,10 @@ class StatsServiceRegisterProvider extends RegisterProvider {
   register(slc) {
     const config = slc.use("Config")
     const sessionService = slc.use("SessionService")
+    const mongoConnection = slc.use("MongoConnection")
+    const redisClient = slc.use("RedisClient")
 
-    return new StatsService(config, sessionService)
+    return new StatsService(config, sessionService, mongoConnection, redisClient)
   }
 }